+++ src/app/globals.css
@@ -8,6 +8,7 @@
   --padding-contact: 8em 24em;
 
   /* z indices */
+  --z-index-eight: 8;
   --z-index-seven: 7;
   --z-index-six: 6;
   --z-index-five: 5;
@@ -66,7 +67,6 @@ body {
   max-width: 100vw;
   height: 100%;
   color: white;
-  background-color: var(--color-article-purple);
 }
 
 a {
