+++ src/components/index.ts
@@ -4,7 +4,7 @@ export { Base } from "./base"
 export { ContactForm } from "./contact-form"
 export { EmailInput } from "./email-input"
 export { Footer } from "./footer"
-export { Header } from "./header"
+export { Header, DynamicHeader } from "./header"
 export { Home } from "./home"
 export { Ticker } from "./ticker"
 export { Puncture } from "./puncture"
@@ -13,7 +13,6 @@ export * from "./Dummy"
 export * from "./GrainBackgrounds"
 export * from "./JupiterLogo"
 export * from "./MenuButton"
-export * from "./NavBar"
 export * from "./NotifyMeGraphic"
 export * from "./XIcon"
 export * from "./StyledComponentsRegistry"
